+++ include/global.h
@@ -5,7 +5,7 @@
  *
  * Global data
  *
- * $Id: global.h 7779 2007-03-03 13:55:42Z pippijn $
+ * $Id: global.h 8027 2007-04-02 10:47:18Z nenolod $
  */
 
 #ifndef _GLOBAL_H
@@ -22,8 +22,8 @@ struct me
   char *desc;                   /* server's description               */
   char *actual;                 /* the reported name of the uplink    */
   char *vhost;                  /* IP we bind outgoing stuff to       */
-  uint16_t recontime;           /* time between reconnection attempts */
-  uint16_t restarttime;         /* time before restarting             */
+  unsigned int recontime;           /* time between reconnection attempts */
+  unsigned int restarttime;         /* time before restarting             */
   char *netname;                /* IRC network name                   */
   char *hidehostsuffix;         /* host suffix for P10 +x etc         */
   char *adminname;              /* SRA's name (for ADMIN)             */
@@ -31,22 +31,22 @@ struct me
   char *mta;                    /* path to mta program                */
   char *numeric;		/* server numeric		      */
 
-  uint32_t loglevel;            /* logging level                      */
-  uint32_t maxfd;               /* how many fds do we have?           */
-  uint32_t mdlimit;		/* metadata entry limit		      */
+  unsigned int loglevel;            /* logging level                      */
+  unsigned int maxfd;               /* how many fds do we have?           */
+  unsigned int mdlimit;		/* metadata entry limit		      */
   time_t start;                 /* starting time                      */
   server_t *me;                 /* pointer to our server struct       */
   boolean_t connected;          /* are we connected?                  */
   boolean_t bursting;           /* are we bursting?                   */
   boolean_t recvsvr;		/* received server peer               */
 
-  uint16_t maxlogins;           /* maximum logins per username        */
-  uint16_t maxusers;            /* maximum usernames from one email   */
-  uint16_t maxnicks;            /* maximum nicks from one username    */
-  uint16_t maxchans;            /* maximum chans from one username    */
-  uint8_t auth;                 /* registration auth type             */
-  uint16_t emaillimit;          /* maximum number of emails sent      */
-  uint16_t emailtime;           /* ... in this amount of time         */
+  unsigned int maxlogins;           /* maximum logins per username        */
+  unsigned int maxusers;            /* maximum usernames from one email   */
+  unsigned int maxnicks;            /* maximum nicks from one username    */
+  unsigned int maxchans;            /* maximum chans from one username    */
+  unsigned int auth;                /* registration auth type             */
+  unsigned int emaillimit;          /* maximum number of emails sent      */
+  unsigned int emailtime;           /* ... in this amount of time         */
 
   time_t uplinkpong;            /* when the uplink last sent a PONG   */
 
@@ -67,19 +67,19 @@ struct ConfOption
 {
   char *chan;                   /* channel we join/msg        */
 
-  uint16_t flood_msgs;          /* messages determining flood */
-  uint16_t flood_time;          /* time determining flood     */
-  uint32_t kline_time;          /* default expire for klines  */
-  uint16_t commit_interval;     /* interval between commits   */
-  int32_t expire;               /* time before registrations expire */
+  unsigned int flood_msgs;          /* messages determining flood */
+  unsigned int flood_time;          /* time determining flood     */
+  unsigned int kline_time;          /* default expire for klines  */
+  unsigned int commit_interval;     /* interval between commits   */
+  int expire;               /* time before registrations expire */
 
   boolean_t silent;             /* stop sending WALLOPS?      */
   boolean_t join_chans;         /* join registered channels?  */
   boolean_t leave_chans;        /* leave channels when empty? */
   boolean_t secure;             /* require /msg <service>@host? */
 
-  uint16_t defuflags;           /* default username flags     */
-  uint16_t defcflags;           /* default channel flags      */
+  unsigned int defuflags;           /* default username flags     */
+  unsigned int defcflags;           /* default channel flags      */
 
   boolean_t raw;                /* enable raw/inject?         */
 
@@ -96,40 +96,40 @@ struct Database
   char *pass;
   char *database;
   char *host;
-  uint32_t port;
+  unsigned int port;
 } database_options;
 
 /* keep track of how many of what we have */
 struct cnt
 {
-  uint32_t event;
-  uint32_t soper;
-  uint32_t svsignore;
-  uint32_t tld;
-  uint32_t kline;
-  uint32_t server;
-  uint32_t user;
-  uint32_t chan;
-  uint32_t chanuser;
-  uint32_t myuser;
-  uint32_t mynick;
-  uint32_t mychan;
-  uint32_t chanacs;
-  uint32_t node;
-  uint32_t bin;
-  uint32_t bout;
-  uint32_t uplink;
-  uint32_t operclass;
-  uint32_t myuser_access;
+  unsigned int event;
+  unsigned int soper;
+  unsigned int svsignore;
+  unsigned int tld;
+  unsigned int kline;
+  unsigned int server;
+  unsigned int user;
+  unsigned int chan;
+  unsigned int chanuser;
+  unsigned int myuser;
+  unsigned int mynick;
+  unsigned int mychan;
+  unsigned int chanacs;
+  unsigned int node;
+  unsigned int bin;
+  unsigned int bout;
+  unsigned int uplink;
+  unsigned int operclass;
+  unsigned int myuser_access;
 };
 
 E struct cnt cnt;
 
 typedef struct claro_state_ {
-	uint32_t node;
-	uint32_t event;
+	unsigned int node;
+	unsigned int event;
 	time_t currtime;
-	uint16_t maxfd;
+	unsigned int maxfd;
 } claro_state_t;
 
 E claro_state_t claro_state;
