+++ src/tokenize.c
@@ -5,16 +5,16 @@
  *
  * This file contains IRC interaction routines.
  *
- * $Id: tokenize.c 7839 2007-03-06 00:09:30Z pippijn $
+ * $Id: tokenize.c 8027 2007-04-02 10:47:18Z nenolod $
  */
 
 #include "atheme.h"
 #include "pmodule.h"
 
-int8_t sjtoken(char *message, char delimiter, char **parv)
+int sjtoken(char *message, char delimiter, char **parv)
 {
 	char *next;
-	uint16_t count;
+	unsigned int count;
 
 	if (!message)
 		return 0;
@@ -72,11 +72,11 @@ int8_t sjtoken(char *message, char delimiter, char **parv)
 }
 
 /* this splits apart a message with origin and command picked off already */
-int8_t tokenize(char *message, char **parv)
+int tokenize(char *message, char **parv)
 {
 	char *pos = NULL;
 	char *next;
-	uint8_t count = 0;
+	unsigned int count = 0;
 
 	if (!message)
 		return -1;
